🐛 Fix php setup

+++ workspaces/custom/workspaces/shared/run/before/php_setup.sh
@@ -1,27 +1,50 @@
 #!/usr/bin/env bash
 set -o errexit -o nounset -o pipefail -o xtrace
 
-# Install Composer
-EXPECTED_SIGNATURE=$(wget -q -O - https://composer.github.io/installer.sig)
-php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-ACTUAL_SIGNATURE=$(php -r "echo hash_file('SHA384', 'composer-setup.php');")
+install_composer () {
+  EXPECTED_SIGNATURE=$(wget -q -O - https://composer.github.io/installer.sig)
+  php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+  ACTUAL_SIGNATURE=$(php -r "echo hash_file('SHA384', 'composer-setup.php');")
 
-if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
-then
-    >&2 echo 'ERROR: Invalid installer signature'
-    rm composer-setup.php
-    exit 1
+  if [ "$EXPECTED_SIGNATURE" != "$ACTUAL_SIGNATURE" ]
+  then
+      >&2 echo 'ERROR: Invalid installer signature'
+      rm composer-setup.php
+      exit 1
+  fi
+
+  php composer-setup.php --install-dir=/usr/local/bin/ --filename=composer
+  rm composer-setup.php
+}
+
+if ! command -v composer >/dev/null; then
+  install_composer
 fi
 
-php composer-setup.php --install-dir=/usr/local/bin/ --filename=composer
-rm composer-setup.php
+put_composer_bins_in_path () {
+  CGR_BINS="$(composer config -g home)/vendor/bin"
+  for i in "$CGR_BINS"/*; do
+    if [[ ! -e "/usr/local/bin/$(basename "$i")" ]]; then
+      ln -s "$i" /usr/local/bin/
+    fi
+  done
+}
 
-composer global require consolidation/cgr
+# Install consolidation/cgr: a safer alternative to `composer global require`
+if ! command -v cgr >/dev/null; then
+  composer global require consolidation/cgr
+  put_composer_bins_in_path
+fi
 
+# Install all other composer packages using cgr
 cgr pantheon-systems/terminus
+cgr drupal/coder
 
-# Link composer installed binaries into part of PATH
-CGR_BINS="$(composer config -g home)/vendor/bin"
-for i in "$CGR_BINS"/*; do
-  ln -s "$i" /usr/local/bin/
-done
+put_composer_bins_in_path
+
+# Register Drupal coding standard with PHPCS tool
+INSTALLED_PATH=~/.composer/global/drupal/coder/vendor/drupal/coder/coder_sniffer
+if (command -v phpcs >/dev/null && [[ -d $INSTALLED_PATH ]]); then
+  phpcs --config-set installed_paths $INSTALLED_PATH
+  phpcs -i
+fi
